requirements.txt
pbr

falcon==0.2.0b2
oslo.db
oslo.log
oslo.utils
sqlalchemy
